requirements.txt
# ML and Computer Vision
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
aiofiles>=23.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
